requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
playwright==1.48.0
pydantic==2.9.0
pydantic-settings==2.5.0
apscheduler==3.10.4
httpx==0.27.0
anthropic>=0.39.0
google-cloud-firestore>=2.16.0
setuptools>=70.0.0
playwright-stealth==1.0.6
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
